Use native C kernel for prior_program
Switch prior_program to call the native C logprior_program from _ckernels.estimation and remove the in-module numba fallback implementations. Add tests/_oracles/estimation.py: a numba-compiled reference/oracle of the packed log-prior kernel relocated from the library. Update parity tests to import and use the oracle (and to call njit-compiled functions directly) so the native C kernels are compared against the numba oracle. Adjusted test imports accordingly to keep bit-for-bit parity checks.
